Update README.md
Please complete the following questions based on the PyTorch tutorial.

https://pytorch.org/docs/stable/nn.html

1.Design a Linear layer using nn. Linear ().

nn.Linear() is used to implement fully connected layer functions. The feature is in B×D format. B denotes batch_size, D denotes the dimension of expected features in the input. Input feature is 32×512. Please design the corresponding layer to output a tensor with 32×256 dimension and print the dimension of the output. Fill in the blanks below to meet the requirements.

input = torch.randn(32, 512)

m = nn.Linear(_________)

output = m(input)